Move replacement tests to a separate file
